Back end/CloudRun/WordCloud/requirements.txt
blis==0.4.1
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.3
Flask==1.1.2
google-api-core==1.22.0
google-auth==1.19.2
google-cloud-core==1.3.0
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.0
googleapis-common-protos==1.52.0
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonify==0.5
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.0
murmurhash==1.0.2
numpy==1.19.1
pandas==1.0.5
Pillow==7.2.0
plac==1.1.3
plotly==4.9.0
preshed==3.0.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
retrying==1.3.3
rsa==4.6
six==1.15.0
spacy==2.3.2
srsly==1.0.2
thinc==7.4.1
tqdm==4.48.0
urllib3==1.25.10
uuid==1.30
wasabi==0.7.1
Werkzeug==1.0.1
wordcloud==1.7.0
zipp==3.1.0
